chore: re-organize react package folder structure. Keep components in components, keep stories in stories.
