+++ packages/react-components/src/components/Badge.jsx
@@ -1,0 +1,7 @@
+import PropTypes from 'prop-types';
+
+export const Badge = ({ label }) => <div className="cbp-badge">{label}</div>;
+
+Badge.propTypes = {
+  label: PropTypes.string,
+};
