--- packages/react-components/src/components/Badge/index.jsx
@@ -1,9 +1,0 @@
-import PropTypes from 'prop-types';
-
-const Badge = ({ label }) => <div className="cbp-badge">{label}</div>;
-
-Badge.propTypes = {
-  label: PropTypes.string,
-};
-
-export default Badge;
