+++ packages/react-components/src/main.js
@@ -1,4 +1,4 @@
 import '@cbpds/vanilla/dist/style.css';
 
-export { default as Badge } from './components/Badge';
-export { default as Button } from './components/Button';
+export { Badge } from './components/Badge';
+export { Button } from './components/Button';
